Task: Geohub harvester refactor.
Refactor Geohub harvest to not refer to a converter (no need to convert, just make package). As well, include business logic for restricted geohub datasets (and licences) and identify non-data resources.
