+++ ckanext/ontario_theme/schemas/external/ontario_theme_dataset.json
@@ -696,6 +696,13 @@
           }
         },
         {
+          "value": "metadata",
+          "label": {
+            "en": "metadata",
+            "fr": "Métadonnées"
+          }
+        },
+        {
           "value": "information",
           "label": {
             "en": "Information",
